adjusted paging in mon plan

+++ common/modules/rpmes/controllers/PlanController.php
@@ -234,14 +234,18 @@ class PlanController extends \yii\web\Controller
         if(Yii::$app->request->get())
         {
             $ids = Yii::$app->request->get('id');
-            $ids = explode(",", $ids);
+            
 
-            if(!empty($ids)){ 
+            if(isset($ids))
+            {
+                $ids = explode(",", $ids);
+                if(!empty($ids)){ 
                     
-                Project::deleteAll(['in', 'id', $ids]);
-                \Yii::$app->getSession()->setFlash('success', 'Selected projects has been deleted successfully'); 
-                return $this->redirect(['/rpmes/plan']);
-            }       
+                    Project::deleteAll(['in', 'id', $ids]);
+                    \Yii::$app->getSession()->setFlash('success', 'Selected projects has been deleted successfully'); 
+                    return $this->redirect(['/rpmes/plan']);
+                }      
+            } 
         }
 
         if(
